+++ src/store.js
@@ -63,7 +63,7 @@ Project*/
 
 // Global stuff ////////////////////////////////////////////////////////
 
-modules.store = '2021-November-11';
+modules.store = '2021-November-12';
 
 // XML_Serializer ///////////////////////////////////////////////////////
 /*
@@ -381,8 +381,8 @@ SnapSerializer.prototype.loadScene = function (xmlNode, remixID) {
     scene.unifiedPalette = model.scene.attributes.palette === 'single';
     scene.showCategories = model.scene.attributes.categories !== 'false';
     scene.disableClickToRun = model.scene.attributes.clickrun === 'false';
-    scene.penColorModel = model.scene.attributes.colormodel === 'hsv' ?
-        'hsv' : 'hsl';
+    scene.penColorModel = model.scene.attributes.colormodel === 'hsl' ?
+        'hsl' : 'hsv';
     model.notes = model.scene.childNamed('notes');
     if (model.notes) {
         scene.notes = model.notes.contents;
@@ -675,8 +675,8 @@ SnapSerializer.prototype.loadBlocks = function (xmlString, targetStage) {
     }
     model.palette = model.childNamed('palette');
     if (model.palette) {
-        SpriteMorph.prototype.customCategories = this.loadPalette(
-            model.palette
+        this.loadPalette(model.palette).forEach((value, key) =>
+            SpriteMorph.prototype.customCategories.set(key, value)
         );
     }
     model.removeChild(model.palette);
@@ -1745,15 +1745,15 @@ Scene.prototype.toXML = function (serializer) {
             '<headers>%</headers>' +
             '<code>%</code>' +
             '<blocks>%</blocks>' +
-            '<variables>%</variables>' +
             '%' + // stage
+            '<variables>%</variables>' +
             '</scene>',
         this.name || localize('Untitled'),
         this.unifiedPalette ? ' palette="single"' : '',
         this.unifiedPalette && !this.showCategories ?
             ' categories="false"' : '',
         this.disableClickToRun ? ' clickrun="false"' : '',
-        this.penColorModel === 'hsv' ? ' colormodel="hsv"' : '',
+        this.penColorModel === 'hsl' ? ' colormodel="hsl"' : '',
         this.notes || '',
         serializer.paletteToXML(this.customCategories),
         Object.keys(this.hiddenPrimitives).reduce(
@@ -1763,8 +1763,8 @@ Scene.prototype.toXML = function (serializer) {
         code('codeHeaders'),
         code('codeMappings'),
         serializer.store(this.stage.globalBlocks),
-        serializer.store(this.globalVariables),
-        serializer.store(this.stage)
+        serializer.store(this.stage),
+        serializer.store(this.globalVariables)
     );
     return xml;
 };